Refactor UpdateScopeFromExecuteAsync
