+++ tracer/src/Datadog.Trace/ClrProfiler/AutoInstrumentation/GraphQL/HotChocolate/ExecuteAsyncIntegrationExtra.cs
@@ -6,6 +6,7 @@
 using System;
 using System.ComponentModel;
 using Datadog.Trace.ClrProfiler.CallTarget;
+using Datadog.Trace.DuckTyping;
 
 namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.GraphQL.HotChocolate
 {
@@ -45,8 +46,24 @@ namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.GraphQL.HotChocolate
         internal static CallTargetState OnMethodBegin<TTarget, TOperationContext>(TTarget instance, TOperationContext operationContext)
             where TOperationContext : IOperationContext
         {
-            HotChocolateCommon.UpdateScopeFromExecuteAsync(Tracer.Instance, operationContext);
+            var operation = operationContext.Operation;
+            var operationType = operation.OperationType.ToString();
+            var operationName = GetNameStringValue(operation.Name.Value);
+
+            HotChocolateCommon.UpdateScopeFromExecuteAsync(Tracer.Instance, operationType, operationName);
             return CallTargetState.GetDefault();
+        }
+
+        private static string GetNameStringValue(object nameValue)
+        {
+            // Get the string value of the NameString
+            // The NameString value can be either another NameString or a string
+            if (nameValue != null && nameValue.TryDuckCast(typeof(NameStringProxy), out var ns))
+            {
+                nameValue = ((NameStringProxy)ns).Value;
+            }
+
+            return nameValue is not null && nameValue is string str ? str : null;
         }
     }
 }
