+++ tracer/src/Datadog.Trace/ClrProfiler/AutoInstrumentation/GraphQL/HotChocolate/ExecuteAsyncIntegrationExtraV13.cs
@@ -36,7 +36,11 @@ namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.GraphQL.HotChocolate
         internal static CallTargetState OnMethodBegin<TTarget, TOperationContext>(TTarget instance, TOperationContext operationContext)
             where TOperationContext : IOperationContextV13
         {
-            HotChocolateCommon.UpdateScopeFromExecuteAsyncV13(Tracer.Instance, operationContext);
+            var operation = operationContext.Operation;
+            var operationType = operation.OperationType.ToString();
+            var operationName = operation.Name;
+
+            HotChocolateCommon.UpdateScopeFromExecuteAsync(Tracer.Instance, operationType, operationName);
             return CallTargetState.GetDefault();
         }
     }
